Make changes per review
