+++ examples/medplum-chart-demo/src/bots/core/gynecology-encounter-note.ts
@@ -1,8 +1,10 @@
 import { BotEvent, getQuestionnaireAnswers, MedplumClient } from '@medplum/core';
 import {
   Bundle,
+  ClinicalImpression,
   CodeableConcept,
   Coding,
+  Condition,
   Encounter,
   Observation,
   Practitioner,
@@ -10,15 +12,16 @@ import {
   Reference,
 } from '@medplum/fhirtypes';
 import {
-  calculateBMI,
   ClinicalImpressionData,
   ConditionData,
-  createClinicalImpressionEntry,
-  createConditionEntries,
-  createObservationEntries,
+  createBundle,
+  createClinicalImpressions,
+  createConditions,
+  createObservations,
   handleBloodPressure,
   ObservationData,
 } from './charting-utils';
+import { calculateBMI } from './observation-utils';
 
 export interface GynecologyObservationData extends ObservationData {
   lastPeriod?: string;
@@ -75,25 +78,20 @@ export async function handler(event: BotEvent<QuestionnaireResponse>, medplum: M
 
   // Create bundle entries from the above objects
   const partialObservations = createPartialGynecologyObservations(observationData, gynecologyCodes);
-  const observationEntries = createObservationEntries(observationData, encounter, user, partialObservations);
-  const conditionEntry = createConditionEntries(conditionData, encounter, user);
-  const clinicalImpressionEntry = createClinicalImpressionEntry(clinicalImpressionData, encounter, user);
+  const observations = createObservations(observationData, encounter, user, partialObservations);
+  const conditions = createConditions(conditionData, encounter, user);
+  const clinicalImpressions = createClinicalImpressions(clinicalImpressionData, encounter, user);
 
   // Create an array of bundle entries for all resource types
-  const entry = [...observationEntries, ...conditionEntry];
-  if (clinicalImpressionEntry) {
-    entry.push(clinicalImpressionEntry);
+  const resources: (Observation | Condition | ClinicalImpression)[] = [...observations, ...conditions];
+  if (clinicalImpressions) {
+    resources.push(clinicalImpressions);
   }
 
-  // Create a batch
-  const resourceBatch: Bundle = {
-    resourceType: 'Bundle',
-    type: 'batch',
-    entry,
-  };
+  const bundle = createBundle(resources);
 
   // Execute the batch to create all resources at once
-  const responseBundle = await medplum.executeBatch(resourceBatch);
+  const responseBundle = await medplum.executeBatch(bundle);
   return responseBundle;
 }
