+++ examples/medplum-chart-demo/src/components/graphs/ObservationGraph.tsx
@@ -1,10 +1,10 @@
 import { Paper } from '@mantine/core';
 import { formatDate, getReferenceString } from '@medplum/core';
-import { Coding, ObservationComponent, Patient } from '@medplum/fhirtypes';
+import { Coding, Observation, ObservationComponent, Patient } from '@medplum/fhirtypes';
 import { useMedplum } from '@medplum/react';
 import { useEffect, useState } from 'react';
 import { LineChart } from './LineChart';
-import { measurementsMeta } from './MeasurementData';
+import { measurementStyles } from './measurement-constants';
 import { ChartData, ChartDataset } from 'chart.js';
 
 interface ObservationGraphProps {
@@ -15,8 +15,11 @@ interface ObservationGraphProps {
 export function ObservationGraph(props: ObservationGraphProps): JSX.Element {
   const medplum = useMedplum();
 
-  const { chartDatasets } = measurementsMeta[props.code.display as string];
+  // Get the chartDataset codes and units
+  const { chartDatasets } = measurementStyles[props.code.display as string];
   const [chartData, setChartData] = useState<ChartData<'line', number[], string>>();
+
+  // Get all of the observations for the current code and patient
   const observations = medplum
     .searchResources('Observation', {
       code: props.code.code,
@@ -27,19 +30,12 @@ export function ObservationGraph(props: ObservationGraphProps): JSX.Element {
   useEffect(() => {
     if (observations) {
       const labels: string[] = [];
+      // For each data item, create an object containing the code and units, as well as a data array to store the measurement values
       const datasets: ChartDataset<'line', number[]>[] = chartDatasets.map((item) => ({ ...item, data: [] }));
 
-      for (const observation of observations) {
-        labels.push(formatDate(observation.effectiveDateTime));
-        if (chartDatasets.length === 1) {
-          datasets[0].data.push(observation.valueQuantity?.value as number);
-        } else {
-          for (let i = 0; i < chartDatasets.length; i++) {
-            datasets[i].data.push((observation.component as ObservationComponent[])[i].valueQuantity?.value as number);
-          }
-        }
-      }
+      getObservationValues(observations, labels, chartDatasets, datasets);
 
+      // Set your chart data with the labels and observation measurments
       setChartData({ labels, datasets });
     }
   }, [chartDatasets, observations]);
@@ -53,4 +49,32 @@ export function ObservationGraph(props: ObservationGraphProps): JSX.Element {
   }
 
   return <div>{chartData && <LineChart chartData={chartData} />}</div>;
+}
+
+function getObservationValues(
+  observations: Observation[],
+  labels: string[],
+  chartDatasets: {
+    label: string;
+    code?: string | undefined;
+    unit: string;
+    backgroundColor: string;
+    borderColor: string;
+  }[],
+  datasets: ChartDataset<'line', number[]>[]
+): void {
+  for (const observation of observations) {
+    // For each observation add the date that it was taken as a label
+    labels.push(formatDate(observation.effectiveDateTime));
+    // If there is only one observation type for the chart, add the observation measurement to your data
+    if (chartDatasets.length === 1) {
+      datasets[0].data.push(observation.valueQuantity?.value as number);
+    } else {
+      // If there is more than one type, loop over each type and add the data. For example, blood pressure may have a systolic and diastolic measurement.
+      // For more details, see https://www.medplum.com/docs/charting/capturing-vital-signs#multi-component-observations
+      for (let i = 0; i < chartDatasets.length; i++) {
+        datasets[i].data.push((observation.component as ObservationComponent[])[i].valueQuantity?.value as number);
+      }
+    }
+  }
 }
